Fix invalid refresh when deleting on last page
If the last page had only one row, and that row was deleted, the table
would be refreshed to an invalid page. This commit changes that
behaviour so that the previous page is loaded instead if a previous page
exists; otherwise the table will simply be empty, but will be on a valid
page.