Prevent race condition when switching pages
This commit introduces the requestState field to RichTableBag so that
consumers have access to whether the table is in the process of making
a request or not. RichTable takes advantage of this to disable
pagination when the table is loading, thus preventing a race condition
where switching pages while the table was loading would result in a
discrepancy between the data displayed and the page number. This commit
also introduces displaying an error message if the table request fails
and improvements to the linear loader animation.