Fix issues in RichTable (#25)
* Fix pagination in RichTable

* Display loading state whenever making a network request

* Replace useCallback with useRef in internalExecuteAction

* Allow axios config to be passed into makeRequest

Also rename the refresh action to make it clear it is an internal action

* Make page navigation an action

* Add count field to RichTableBag

Also used the new count field to show the displayed range in the RichTable
footer.

* Fix invalid refresh when deleting on last page

If the last page had only one row, and that row was deleted, the table
would be refreshed to an invalid page. This commit changes that
behaviour so that the previous page is loaded instead if a previous page
exists; otherwise the table will simply be empty, but will be on a valid
page.

* Update react-bootstrap

* Add page selector to RichTable pagination

* Make RichTable responsive

This also means that long table cells wrap their contents, which lets
us modify spacing in UserList.

* Revert "Update react-bootstrap"

This reverts commit ef9383b4. See #27.

* Add linear loader to rich tables

* Only show one loader to screenreaders

* add newline to Loader

* Prevent race condition when switching pages

This commit introduces the requestState field to RichTableBag so that
consumers have access to whether the table is in the process of making
a request or not. RichTable takes advantage of this to disable
pagination when the table is loading, thus preventing a race condition
where switching pages while the table was loading would result in a
discrepancy between the data displayed and the page number. This commit
also introduces displaying an error message if the table request fails
and improvements to the linear loader animation.

* Simplify page input and correct Loader props

- Use the current page number as an input instead of displaying a
  completely new component.
- Change page input behaviour so that a blur that is the result of
  clicking outside of the input submits the new page. All other blurs
  are treated as a cancel.
- Rename style prop of Loader to variant for consistency with Bootstrap
  and to avoid conflict with default style prop.

* Split header and footer into separate components

* Fix small tables not filling container on mobile

* Fix errors persisting after successful responses in useApiGet

* Run prettier for compatibility with #33

* Don't validate page input until submit